Package refactoring for enums
