+++ core/src/main/java/bg/tu/varna/events/core/processors/user/RegisterUserProcessor.java
@@ -5,8 +5,8 @@ import bg.tu.varna.events.api.exceptions.UserExistsException;
 import bg.tu.varna.events.api.operations.user.register.personal.RegisterUserOperation;
 import bg.tu.varna.events.api.operations.user.register.personal.RegisterUserRequest;
 import bg.tu.varna.events.api.operations.user.register.personal.RegisterUserResponse;
-import bg.tu.varna.events.persistence.entities.Role;
 import bg.tu.varna.events.persistence.entities.User;
+import bg.tu.varna.events.persistence.enums.Role;
 import bg.tu.varna.events.persistence.repositories.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
